requirements.txt
Flask
python-dotenv
requests
symbl